+++ NEWS.md
@@ -5,6 +5,7 @@
 * new function `plot_categorical()` (#47)
 * new colour palette `safe_pal()` and corresponding scales `scale_fill_safe()`
   and `scale_colour_safe()` for `ggplot2`
+* new argument `minimal` for `plot_proportions()`
 
 **Minor breaking change**